# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "빵시계 API 서버입니다.",
        "title": "빵시계 API 서버",
        "contact": {},
        "version": "1.0"
    },
    "host": "breadclock.hyunchung.dev",
    "basePath": "/api/v1",
    "paths": {
        "/auth/login/:provider": {
            "post": {
                "description": "OAuth2로 발급 받은 코드를 이용한 토큰 발급 및 로그인 처리",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "로그인",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OAuth2 provider",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "로그인 요청 정보",
                        "name": "loginRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.loginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/bakeries": {
            "get": {
                "description": "전체 빵집 목록 조회 (sort param 미지정 시 이름순 정렬)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bakeries"
                ],
                "summary": "빵집 목록 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "정렬 옵션 (name|distance)",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "조회 개수",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "조회 offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "필터 옵션 (favorites)",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "현재 위치 좌표값 (위도,경도)",
                        "name": "loc",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.BakeryList"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/bakeries/:bakeryId": {
            "get": {
                "description": "특정 빵집에 대한 상세 정보 조회",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bakeries"
                ],
                "summary": "빵집 상세 조회",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "빵집 ID",
                        "name": "bakeryId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "현재 위치 좌표값 (위도,경도)",
                        "name": "loc",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.BakeryDetail"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/bakeries/:bakeryId/breads/availability": {
            "put": {
                "description": "특정 빵집의 빵 종류별 매진 정보 갱신",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bakeries"
                ],
                "summary": "빵 매진 정보 갱신",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "빵집 ID",
                        "name": "bakeryId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "빵 정보 리스트",
                        "name": "BreadList",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.BreadList"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/bakeries/:bakeryId/favorite": {
            "put": {
                "description": "특정 빵집을 즐겨찾기에 추가",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bakeries"
                ],
                "summary": "빵집 즐겨찾기 추가",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "빵집 ID",
                        "name": "bakeryId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "description": "특정 빵집을 즐겨찾기에서 해제",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bakeries"
                ],
                "summary": "빵집 즐겨찾기 해제",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "빵집 ID",
                        "name": "bakeryId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/search": {
            "get": {
                "description": "특정 빵 종류가 제공되는 빵집들을 검색 (sort param 미지정 시 이름순 정렬)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Search"
                ],
                "summary": "빵 이름으로 빵집 검색",
                "parameters": [
                    {
                        "type": "string",
                        "description": "검색 문자열",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "정렬 옵션 (name|distance)",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "조회 개수",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "조회 offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "필터 옵션 (favorites)",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "현재 위치 좌표값 (위도,경도)",
                        "name": "loc",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.BakeryList"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "api.Bakery": {
            "type": "object",
            "properties": {
                "breads": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.BreadDetail"
                    }
                },
                "coordinates": {
                    "type": "string"
                },
                "favorite": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "photoUrls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "api.BakeryDetail": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "breads": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.BreadDetail"
                    }
                },
                "coordinates": {
                    "type": "string"
                },
                "favorite": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "openingHours": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.OpeningHours"
                    }
                },
                "photoUrls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "api.BakeryList": {
            "type": "object",
            "properties": {
                "bakeries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.Bakery"
                    }
                }
            }
        },
        "api.Bread": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "api.BreadDetail": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "boolean"
                },
                "availableHours": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "photoUrl": {
                    "type": "string"
                }
            }
        },
        "api.BreadList": {
            "type": "object",
            "properties": {
                "breads": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.Bread"
                    }
                }
            }
        },
        "api.OpeningHours": {
            "type": "object",
            "properties": {
                "close": {
                    "type": "string"
                },
                "open": {
                    "type": "string"
                }
            }
        },
        "api.loginRequest": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                }
            }
        }
    }
}